adopt sources after renaming files
